+++ biotransformers/utils/constant.py
@@ -6,7 +6,8 @@ ESM_LIST = [
     "esm1_t6_43M_UR50S",
     "esm1b_t33_650M_UR50S",
     "esm_msa1_t12_100M_UR50S",
-    "esm1v_t33_650M_UR90S_1"
+    "esm_msa1b_t12_100M_UR50S",
+    "esm1v_t33_650M_UR90S_1",
 ]
 
 ROSTLAB_LIST = ["Rostlab/prot_bert", "Rostlab/prot_bert_bfd"]